refactor(agent): split broker_manager.rs into broker_manager/ submodules
